refactor(testing): expose recording HTTP client
